+++ packages/article/article-tracking-context.js
@@ -4,7 +4,6 @@ import { withTrackingContext } from "@times-components/tracking";
 export default Component =>
   withTrackingContext(Component, {
     trackingObjectName: "Article",
-    isDataReady: ({ isLoading }) => !isLoading,
     getAttrs: ({ article } = {}) => ({
       byline: get(article, "byline[0].children[0].attributes.value", ""),
       headline: get(article, "headline", ""),