+++ packages/slice/__tests__/ios/jest.config.js
@@ -1,3 +1,7 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = jestConfigurator("ios", __dirname, ["responsive.js"]);
+module.exports = jestConfigurator("ios", __dirname, [
+  "responsive.web.js",
+  "**/styles",
+  "slice.js"
+]);