+++ packages/slice/slice.js
@@ -1,4 +1,3 @@
-// @TODO: delete this file when related articles wired up properly
 import React from "react";
 import { StyleSheet, View } from "react-native";
 import { propTypes, defaultProps } from "./proptypes";