+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.10"></a>
+## [0.8.10](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.8.9...@times-components/error-view@0.8.10) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.8.9"></a>
 ## [0.8.9](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.8.8...@times-components/error-view@0.8.9) (2018-02-28)